Rename project
There's already a project named parakeet on PyPI, so renaming to avoid
the conflict.
